Added important notes about puplisher concept.
and how to act and send values to subscribers.
It's very important to know that handlers work form two sides, the listeners and publisher, the listener is the guy who will get notified when an event will be emitted or sent, the publisher is the guy who will send/emit these values.
In the world of closures or handlers emission can happen by keeping a reference to it (normally arrays) and call this item from the array with the value you want, i.e deletions[index](error) or insertions[index](nil)
By understating that the closures work from two sides not only one, you can listen as you like, and also you now know how to push values there by keeping a reference to that handler in array, making the public API insert when its getting called and fetch and execute with the helper testing methods.

Also its very important to define your own completion handler signature, and this can come naturally and easy from the use case -> scenarios -> narrative.

Signed-off-by: Ahmad Atef <ahmad.atef.ali.ahmad@gmail.com>